+++ src/avt/IActionGoKhoHang1.java
@@ -9,6 +9,6 @@ final class IActionGoKhoHang1 implements IAction {
 
    public final void perform() {
       Item var1 = (Item)FarmScr.itemProduct.elementAt(this.a);
-      FarmScr.gI().doSellProduct((int)var1.ID, var1.name);
+      FarmScr.gI().doSellProduct(var1.ID, var1.name);
    }
 }
