+++ src/avt/ImageInfo.java
@@ -15,6 +15,6 @@ public final class ImageInfo {
       int var10003 = this.y0 * AvMain.hd;
       int var10004 = this.w * AvMain.hd;
       int var10005 = this.h * AvMain.hd;
-      var1.drawRegion(AvatarData.getBigImgInfo((int)this.bigID).img, var10002, var10003, var10004, var10005, 0, var2, var3, var4);
+      var1.drawRegion(AvatarData.getBigImgInfo(this.bigID).img, var10002, var10003, var10004, var10005, 0, var2, var3, var4);
    }
 }
